Fix hrtime(true) to return float per php-src (#12779) (#12782)
hrtime(as_number: true) now returns double like microtime(true), matching
ext/standard/hrtime.c. VM, JIT lowering, and HrtimeJitHelper::nsFloat share
the PHP SSOT path.

+++ ext/standard/hrtime.php
@@ -40,7 +40,7 @@ final class hrtime extends Internal
             );
         }
         if ($asNumber) {
-            $frame->returnVar->int(VmDate::hrtime(true));
+            $frame->returnVar->float((float) VmDate::hrtime(true));
 
             return;
         }
